Avoid recursive "unmodifiable" wrappings
Java does not detect if the argument to unmodifiableMap or
unmodifiableSet is already unmodifiable and happily wraps it again. Do
that a lot and you have a problem.

This commit makes sure that TableauNode does not cause such recursive
wrappings. This improves the runtime for model checking some relatively
complicated example from about 8.5 seconds to 2.7 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>