PositiveFormFormulaTransformer: Add support for let formulas
This also has the side-effect of eliminating unused lets. This is
because this code has to do the following:

For the input !(let X = true in X || !X), we want to move all negations
as far into the formula as possible. This means that the negation has to
be pushed into the expansion of the let, but since X appears both
positive and negative, this means that we need a new variable.

Thus, the positive form of the above input would be (let Y = true in let
X = false in X || Y). Here, a new, fresh variable Y was introduced to
express the negated let term. This variable's let term contains the
"original" let term, because since the variable is fresh, it cannot
accidentally be bound.

Since this code should only introduce these new lets when needed, it
becomes very easy to also eliminate unused lets.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>