SolveEquationSystem: Produce nicer formulas
This partially reverts commit 47a86433576e3ae1 (especially its changes
to the tests): If a variable only appears once as a free variable, then
this code now again substitutes the variable directly instead of
introducing a let-expression to do so. Only when a variable occurs
multiple times as a free variable does this generate a let expression.

On the tests: As mentioned above 47a86433576e3ae1 was reverted on the
tests. Additionally, a new test was added where all the "false"
subformulas of the test were replaced with variables so that
let-expressions are introduced by this code.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>