Do not allow let formulas in SubstitutionTransformer
This commit adds a (disabled) test showing that a substitution cannot
really handle let formulas: The free variables inside the expansion part
of the let could become bound in the position where the expansion is
expanded. Worse, some expansion places could bind it while others do
not.

Thus, simply do not allow Let formulas in the input to
SubstitutionTransformer.

This affects SolveEquationSystem, because it can no longer mix
substitutions and Lets. Instead, it now always uses Lets. This in turn
means that some tests are affected since we basically revert commits
49112622e0b94d582b796c and  4949f49c222c0205046.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>