MTSToFormula: Add special mode for deterministic systems
If we know that the formula will be interpreted on a deterministic
system, we can generate simpler formulas. This is because for a may arc
with label 'a' to state 's' the formula <a>s is enough to make sure
things behave correctly. In a non-deterministic system there could be
other 'a'-labelled arcs that we also need to capture via an universal
modality. This is not the case in deterministic systems.

Thus, the resulting formulas will end up being a lot simpler.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>