Heuristic: Optimise TableauBuilder's working order
A disjunction means that the current state of the TableauBuilder has to
be split and each part continues independently. Thus, all
"non-splitting" work that is still pending at this part is done by both
TableauBuilder which exist after the split.

To avoid this duplicate work, the list of pending jobs is now "a bit
ordered": Disjunctions are added to the tail of the list while
everything else is placed at the front of the list.

This also means that this code switches "a bit" from a breadth-first
search order to a depth-first search order, except that disjunctions are
still breadth-first...

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>