Add a new HidingOperatorModule
This module expands the hiding operator that I defined. That means, all
modalities inside the application of the hiding operator are re-written
so that events that do not appear inside of the formula in the hiding
operator are locally ignored.

For example, <a>true || h([b]false) is rewritten into <a>true || nu X.
[a]X && [b]false.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>