Rearrange steps in RealiseFormula.Worker
Previously, the worker first added missing arcs and then did an
overapproximation. Now it first does the overapproximation and then adds
missing arcs. This avoids throwing away the result of continueTableau()
if multiple tableaus are produced. Instead, this needlessly
over-approximates the "empty TS" once, but this should be fast and does
not really hurt at all.

This improves the run time for finding a 9-bounded realisation of 'nu
X.<a><b><c>true&&<b><a>[c]X' from 15 to 9 seconds.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>