TableauBuilder: Evaluate closed formulas only once per state
A closed formula does not depend on anything for its evaluation,
especially not the definition list. Thus, if two different paths cause
the same closed formula to be evaluated for the same state, then it only
needs to be evaluated once.

This commit makes the TableauBuilder track the evaluation of closed
formulas in the Tableau and uses this as an optimisation.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>