Add a specialised variant for realising deadlock-free implementations
The problem with forbidding deadlocks in the formula is that this can
lead to state space explosion. When the over-approximation adds \ell new
states to the lts and the alphabet has size n, then there are n^\ell
ways to resolve the deadlocks.

Since it is not uncommon for the over-approximation to add many states,
this leads to problems. For example, in my attempts to specify dining
philosphers, there is a case where 10 new states appear, which mean that
there are 6^10 ways to forbid deadlocks here, which means that there are
6^10 tableaus to consider in the next step.

So, instead of forbidding deadlocks via the formula, this module takes a
different approach.

The formula is realised as usual, allowing deadlocks. If a deadlock-free
realisation is found, then it is printed. However, if there is a
deadlock, then one deadlock state is picked and all ways to eliminate
this deadlock are generated. The resulting ts are then further realised.

This avoids state explosion by only eliminating one deadlock at a time.
Most of the possibilities that are generated here will turn out to be
inconsistent with the formula in the next step and are discard.

Another plus point is that this does not choose "badly". If the minimal
over-approximation added a state and also added an outgoing edge to this
state, then this state already is not a deadlock. Still, the old
formula-based approach could guess badly and still generate all other
possibilities to "eliminate" the deadlock.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>